today update

+++ controller/qnaController.js
@@ -1,10 +1,14 @@
+import mongoose from "mongoose"
 import { qna } from "../model/qnaModel.js"
 import { ErrorHandler } from "../utils/Errorhandler.js"
 
-export const createQNA = async(req, res, next)=>{
+export const createQNA = async (req, res, next) => {
     try {
-        const {question , answer} = req.body
+        const { question, answer } = req.body
 
+        if (!question || !answer) {
+            return next(new ErrorHandler("Question and Answer are required!", 400));
+        }
 
         const data = await qna.create({
             question,
@@ -14,8 +18,8 @@ export const createQNA = async(req, res, next)=>{
         await data.save()
 
         res.status(200).json({
-            success : true,
-            message : "qna created !",
+            success: true,
+            message: "qna created !",
             data
         })
     } catch (err) {
@@ -23,11 +27,11 @@ export const createQNA = async(req, res, next)=>{
     }
 }
 
-export const getQNA = async(req, res, next)=>{
+export const getQNA = async (req, res, next) => {
     try {
-        const data =  await qna.find()
+        const data = await qna.find()
         res.status(200).json({
-            success : true,
+            success: true,
             data
         })
     } catch (err) {
@@ -36,13 +40,20 @@ export const getQNA = async(req, res, next)=>{
 }
 
 
-export const getOneQNA = async(req, res, next)=>{
+export const getOneQNA = async (req, res, next) => {
     try {
 
         const qnaId = req.params.id
-        const data =  await qna.findOne({_id : qnaId})
+        if (!mongoose.Types.ObjectId.isValid(qnaId)) {
+            return next(new ErrorHandler("Invalid QNA ID!", 400));
+        }
+        const data = await qna.findOne({ _id: qnaId })
+
+        if (!data) {
+            return next(new ErrorHandler("QNA not found!", 404));
+        }
         res.status(200).json({
-            success : true,
+            success: true,
             data
         })
     } catch (err) {
@@ -51,23 +62,29 @@ export const getOneQNA = async(req, res, next)=>{
 }
 
 
-export const updateQNA = async(req, res, next)=>{
+export const updateQNA = async (req, res, next) => {
     try {
 
         const qnaId = req.params.id
-        const {question , answer} = req.body
-        return next(new ErrorHandler("at least one field is requied", 400))
-    const data =  await qna.findOne({_id : qnaId})
+        const { question, answer } = req.body
+        if (!mongoose.Types.ObjectId.isValid(qnaId)) {
+            return next(new ErrorHandler("Invalid QNA ID!", 400));
+        }
+        if (!question && !answer) {
+            return next(new ErrorHandler("At least one field is required!", 400));
+        }
+        const data = await qna.findOne({ _id: qnaId })
 
-        return next(new ErrorHandler("qna not found", 404))
+        if (!data) {
+            return next(new ErrorHandler("QNA not found!", 404));
+        }
+        if (question) data.question = question
+        if (answer) data.answer = answer
 
-    if(question) data.question = question
-    if(answer) data.answer = answer
-
-    await data.save()
+        await data.save()
 
         res.status(200).json({
-            success : true,
+            success: true,
             data
         })
     } catch (err) {
@@ -77,18 +94,24 @@ export const updateQNA = async(req, res, next)=>{
 
 
 
-export const deleteQNA = async(req, res, next)=>{
+export const deleteQNA = async (req, res, next) => {
     try {
 
+
         const qnaId = req.params.id
-        const data =  await qna.findOne({_id : qnaId})
 
-            return next(new ErrorHandler("qna already deleted !", 200))
+        if (!mongoose.Types.ObjectId.isValid(qnaId)) {
+            return next(new ErrorHandler("Invalid QNA ID!", 400));
+        }
+        const data = await qna.findOne({ _id: qnaId })
 
+        if (!data) {
+            return next(new ErrorHandler("QNA already deleted or not found!", 404));
+        }
         await data.deleteOne()
         res.status(200).json({
-            success : true,
-            message : "qna deleted succesfully !"
+            success: true,
+            message: "qna deleted succesfully !"
         })
     } catch (err) {
         return next(new ErrorHandler(err.message, 500))
